// _src/tsBase/svc/EditMany.d.ts
export default class EditMany {
    private mode;
    private modeData;
    private systemError;
    private kid;
    private rowFilter;
    private sortFid;
    private hasRowTpl;
    private hasRowFilter;
    private rowTpl;
    private hasEform;
    private rowsBox;
    private eform;
    private deletedRows;
    private newIndex;
    private validator;
    private fnReset;
    private fnLoadRows;
    private fnValid;
    private fnGetUpdJson;
    _childs: EditMany[];
    fidTypes: string[];
    fidTypeLen: number;
    fidRadios: string[];
    hasFile: boolean;
    fileLen: number;
    fileFids: string[];
    constructor(kid: string, rowsBoxId?: string, rowTplId?: string, rowFilter?: string, sortFid?: string);
    private showErrors;
    private setChilds;
    private initUrm;
    private _isNewBox;
    private reset;
    private _resetVar;
    private _resetDeletes;
    private _urmLoadRows;
    private _urmGetUpdJson;
    private _urmReset;
    private loadRowsBySys;
    private loadRowByBox;
    private loadRowsByRsb;
    private valid;
    private getKey;
    private _checkRowFilter;
    private _checkRowTpl;
    private _elmToRowBox;
    private idToRowBox;
    private getUpdJsonBySys;
    private getUpdJsonByRsb;
    private getUpdRow;
    private getUpdRows;
    private getDeletes;
    private onAddRow;
    private addRow;
    private onDeleteRow;
    private deleteRow;
    private deleteAll;
    private onViewFile;
    private _renderRow;
    private dataAddFiles;
    private rowSetFkey;
    private rowsSetFkey;
    private rowsToNew;
    private setNewIndex;
    private setNewIdByBox;
    private setSort;
    private _getRowsBox;
}